Add voting list
